Deji

+++ Csharp/RT_CS/RT_CS/Cameras/PerspectiveCamera.cs
@@ -7,7 +7,7 @@ namespace RT_CS.Cameras
 {
     public class PerspectiveCamera : Camera
     {
-        public override Ray Project(float px, float py)
+        public override Ray Project(int px, int py)
         {
             Ray myRay;
             Vector3 dir = ComputePixel(px, py);
@@ -16,7 +16,7 @@ namespace RT_CS.Cameras
             return myRay;
 
         }
-        public PerspectiveCamera (float fl, float fp, Vector3 up, Vector3 po, Vector3 ta, float fi, float scw, float sch) : 
+        public PerspectiveCamera (float fl, float fp, Vector3 up, Vector3 po, Vector3 ta, float fi, int scw, int sch) : 
             base (fl,fp,up,po,ta,fi,scw,sch)
         {
